[PDFBOX - Refactoring] Move Method: method getNodeValue moved from XMLUtil to FDFField as it was being exclusively used in the latter. -- Moving the method makes the FDFField.java class more coherent. -- Moving the method also eliminates the dependency between classes.
